Improve flushing
This flushes after each message, including two non-loop messages I
had missed. It extracts the printing and flushing code, and the
explanation for why I am flushing, to a helper function lprintf,
and calls that instead of printf. The explanation is also improved.

Although I don't want to do anything more complicated than this, it
might be justified to do less, or there may be a better approach.