Make it a bash script to fix job control on CI
Shell job control is mainly for interactive use, and shells vary in
the situations where it is supported. In the noninteractive session
on CI, the "set -m" command in the run-experiment script failed on
Ubuntu, where /bin/sh is dash, but succeeded on macOS, where
/bin/sh is (I think) bash 3.2. This change makes it work on both.

Besides its somewhat unusual use of job control, this script
otherwise currently relies on no "bash-isms".

(SSHing in for tmate was never affected by this, but the main point
of CI is noninteractive use, so I want that to work on both
systems.)