Try to pull complex logic out of step-name interpolation
This would also have the benefit of stopping using the
"a && b || c" fake ternary conditional, which is brittle.

The issue here is that I don't know what variables I can define
that I can interpolate into a *step name*. But using a CI matrix
variable, as here, does not seem to be workable.