Just always have the parenthesized note on the step name (#10)
* Try to pull complex logic out of step-name interpolation

This would also have the benefit of stopping using the
"a && b || c" fake ternary conditional, which is brittle.

The issue here is that I don't know what variables I can define
that I can interpolate into a *step name*. But using a CI matrix
variable, as here, does not seem to be workable.

* Just always have the parenthesized note on the step name

The complex logic in the interpolated name was very bad, both in
terms of the workflow's clarity and maintainability, and in that it
used the brittle fake ternary conditional a && b || c. I had only
introduced that a stopgap while trying to figure out a better way
to express it. I didn't, and it's better not to have the feature
that only provides the parenthesized note when needed, than to have
it but write it this way.